pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0", "wheel"]

[project]
authors = [
  {name = "Zhenzhi Tan", email = "zhenzhi-tan@outlook.com"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "numpy==1.22.4",
  "pandas>=1.3.0",
  "scikit-learn>=1.0.0",
  "scipy>=1.7.0",
  "matplotlib>=3.5.0",
  "seaborn>=0.11.0",
  "joblib>=1.1.0",
  "tqdm>=4.62.0",
  "loguru>=0.6.0",
  "hyperopt>=0.2.7",
  "umap-learn>=0.5.0",
  "lightgbm>=3.3.0",
  "xgboost==1.5.2",
  "catboost>=1.0.0",
  "torch>=1.10.0",
  "shap==0.44.0",
  "rdkit>=2025.3.3",
  "torchbnn>=1.2",
  "openpyxl>=3.1.5",
  "tables>=3.1.0",
]
description = "SynAD: Synthesis Applicability Domain for chemical reaction prediction"
keywords = ["chemistry", "machine learning", "synthesis", "applicability domain", "chemical reactions"]
license = {text = "MIT"}
maintainers = [
  {name = "Zhenzhi Tan", email = "zhenzhi-tan@outlook.com"},
]
name = "synad"
readme = "README.md"
requires-python = ">=3.8"
version = "0.1.0"

[project.optional-dependencies]
all = [
  "synad[dev,jupyter,docs]",
]
dev = [
  "pytest>=6.0",
  "pytest-cov>=2.0",
  "black>=22.0",
  "isort>=5.0",
  "flake8>=4.0",
  "mypy>=0.900",
  "pre-commit>=2.15.0",
]
docs = [
  "sphinx>=4.0.0",
  "sphinx-rtd-theme>=1.0.0",
  "sphinx-autodoc-typehints>=1.12.0",
]
jupyter = [
  "jupyter>=1.0.0",
  "ipywidgets>=7.6.0",
  "ipykernel>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/deepsynthesis/synad"
Repository = "https://github.com/deepsynthesis/synad"

[project.scripts]
synad-evaluate = "synad.cli:evaluate_command"
synad-score = "synad.cli:score_command"
synad-train = "synad.cli:train_command"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
synad = ["data/*.csv", "data/*.json", "models/*.pkl", "configs/*.yaml"]

[tool.black]
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
include = '\.pyi?$'
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
profile = "black"
use_parentheses = true

[tool.flake8]
exclude = [
  ".git",
  "__pycache__",
  "docs/source/conf.py",
  "old",
  "build",
  "dist",
]
extend-ignore = ["E203", "W503", "E501"]
max-line-length = 120

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
no_implicit_optional = true
python_version = "3.8"
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]
minversion = "6.0"
testpaths = [
  "tests",
]

[tool.coverage.run]
omit = [
  "*/tests/*",
  "*/test_*.py",
]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
]
